don't leak passed file descriptors on failure

+++ lib/irq.h
@@ -41,7 +41,7 @@ handle_device_get_irq_info(vfu_ctx_t *vfu_ctx, uint32_t size,
                            struct vfio_irq_info *irq_info_out);
 int
 handle_device_set_irqs(vfu_ctx_t *vfu_ctx, uint32_t size,
-                       int *fds, int nr_fds, struct vfio_irq_set *irq_set);
+                       int *fds, size_t nr_fds, struct vfio_irq_set *irq_set);
 
 #endif /* LIB_VFIO_USER_IRQ_H */
